refactor(vscode): unify webview hosts, onboarding, and extension layout
Extract shared host/SSE/message delivery, split extension activate into
commands/status/sidebar modules, carve webview local API routes, and
wire clear CLI-missing onboarding with walkthrough + settings sync.
