+++ packages/vscode/src/opencode/manager.external-url.test.ts
@@ -1,79 +1,55 @@
-// @ts-nocheck
-import { afterEach, beforeEach, describe, mock, test } from 'bun:test';
+import { describe, test } from 'node:test';
 import assert from 'node:assert/strict';
+import { connectExternalOpenCodeUrl } from './external-url';
+import type { ReadyResult } from './types';
 
-const originalFetch = globalThis.fetch;
-
-const getConfiguration = mock(() => ({
-  get: (key: string) => {
-    if (key === 'apiUrl') return 'http://127.0.0.1:5555/';
-    if (key === 'opencodeBinary') return '';
-    return '';
-  },
-}));
-
-mock.module('vscode', () => ({
-  l10n: { t: (value: string) => value },
-  workspace: {
-    get workspaceFolders() {
-      return [{ uri: { fsPath: '/workspace' } }];
-    },
-    getConfiguration: getConfiguration,
-  },
-  window: {
-    createOutputChannel: () => ({
-      appendLine: () => {},
-    }),
-    showErrorMessage: async () => undefined,
-  },
-  Uri: { parse: (value: string) => ({ toString: () => value }) },
-  env: { openExternal: async () => true },
-  commands: { executeCommand: async () => undefined },
-  Disposable: class {
-    constructor(private readonly disposeFn: () => void) {}
-    dispose() {
-      this.disposeFn();
-    }
-  },
-}));
-
-const { createOpenCodeManager } = await import('./manager');
-
-describe('external configured apiUrl startup', () => {
-  beforeEach(() => {
-    getConfiguration.mockClear();
-  });
-
-  afterEach(() => {
-    globalThis.fetch = originalFetch;
-  });
-
-  test('probes health before marking connected', async () => {
-    let healthChecks = 0;
-    globalThis.fetch = (async (input: Parameters<typeof fetch>[0]) => {
-      const url = String(input);
-      if (url.includes('/global/health')) {
-        healthChecks += 1;
-        return new Response(JSON.stringify({ healthy: true, version: '1.18.8' }), { status: 200 });
-      }
-      return new Response('{}', { status: 404 });
-    }) as typeof fetch;
-
-    const context = {
-      globalStorageUri: { fsPath: '/tmp/openchamber-global-storage' },
+describe('connectExternalOpenCodeUrl', () => {
+  test('marks connected when ready check succeeds', async () => {
+    let calls = 0;
+    const readyCheck = async (): Promise<ReadyResult> => {
+      calls += 1;
+      return {
+        ok: true,
+        baseUrl: 'http://127.0.0.1:5555',
+        elapsedMs: 12,
+        attempts: 1,
+        version: '1.18.8',
+      };
     };
 
-    const manager = createOpenCodeManager(context);
-    const statuses: string[] = [];
-    manager.onStatusChange((status) => {
-      statuses.push(status);
+    const result = await connectExternalOpenCodeUrl(
+      'http://127.0.0.1:5555/',
+      {},
+      5000,
+      readyCheck,
+    );
+
+    assert.equal(calls, 1);
+    assert.equal(result.ok, true);
+    if (result.ok) {
+      assert.equal(result.version, '1.18.8');
+      assert.equal(result.detectedPort, 5555);
+    }
+  });
+
+  test('returns error when health check fails', async () => {
+    const readyCheck = async (): Promise<ReadyResult> => ({
+      ok: false,
+      elapsedMs: 300,
+      attempts: 3,
+      version: null,
     });
 
-    await manager.start();
+    const result = await connectExternalOpenCodeUrl(
+      'http://127.0.0.1:5555/',
+      {},
+      300,
+      readyCheck,
+    );
 
-    assert.equal(healthChecks > 0, true);
-    assert.equal(manager.getStatus(), 'connected');
-    assert.ok(statuses.includes('connecting'));
-    assert.equal(statuses.at(-1), 'connected');
+    assert.equal(result.ok, false);
+    if (!result.ok) {
+      assert.match(result.error, /not healthy/);
+    }
   });
 });
